workers/notify_worker/requirements.txt
aiohttp>=3.9.0
apscheduler>=3.10.4
redis[async]>=5.0.0
motor>=3.4.0
python-telegram-bot>=20.0
pymongo>=4.6.0
python-dotenv>=1.0.0 